refactor: remove legacy components
