--- src/components/Dialog/index.ts
@@ -1,2 +1,0 @@
-import Dialog from "./Dialog";
-export default Dialog;
