--- src/components/InformationDialog/index.ts
@@ -1,2 +1,0 @@
-import InformationDialog from "./InformationDialog";
-export default InformationDialog;
