+++ src/components/index.ts
@@ -1,12 +1,9 @@
 export { default as GlobalStyles } from "./GlobalStyles";
 export { default as Header } from "./Header";
-export { default as Dialog } from "./Dialog";
 export { default as SuperHeader } from "./SuperHeader";
-export { default as Stepper } from "./Stepper";
 export { default as Banner } from "./Banner";
 export { default as Sidebar } from "./Sidebar";
 export { default as LayoutV2 } from "./LayoutV2";
-export { default as DotStepper } from "./DotStepper";
 export { default as Modal } from "./Modal";
 export { default as Alert } from "./Alert";
 export { default as Selector } from "./Selector";
@@ -14,7 +11,6 @@ export { default as InputWithMaxButton } from "./InputWithMaxButton";
 export { Text } from "./Text";
 export { WrongNetworkHeader } from "./WrongNetworkHeader";
 
-export * from "./Box";
 export * from "./Buttons";
 export * from "./ExternalLink";
 export * from "./ErrorBoundary";
